pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "autombse"
version = "0.1.0"
description = "AutoMBSE migration scaffolding and CLI"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "PyYAML>=6.0",
  "jsonschema>=4.0",
  "openai>=1.0",
  "requests>=2.0",
  "tqdm>=4.0",
  "numpy>=1.20",
  "scipy>=1.8",
]

[project.optional-dependencies]
rag = ["torch==2.2.2+cpu", "transformers", "qdrant-client"]
ingest = ["pandas", "torch==2.2.2+cpu", "transformers", "qdrant-client"]
all = ["pandas", "torch", "transformers", "qdrant-client"]

[project.scripts]
autombse = "autombse.cli.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["autombse*"]
